Provide working mySQL query retrieving one entry
readOne function in models, move db connection to database/models/index.js
